+++ web/src/input_pill.ts
@@ -13,18 +13,21 @@ import * as ui_util from "./ui_util";
 // See https://zulip.readthedocs.io/en/latest/subsystems/input-pills.html
 
 export type InputPillItem<T> = {
+    group_id?: number;
     display_value: string;
     type: string;
     img_src?: string;
     deactivated?: boolean;
     status_emoji_info?: EmojiRenderingDetails & {emoji_alt_code?: boolean}; // TODO: Move this in user_status.js
     should_add_guest_user_indicator?: boolean;
+    group_size?: number;
 } & T;
 
 type InputPillCreateOptions<T> = {
     $container: JQuery;
     pill_config?: {
         show_user_status_emoji?: boolean;
+        show_user_group_size?: boolean;
     };
     create_item_from_text: (
         text: string,
@@ -51,6 +54,7 @@ type InputPillStore<T> = {
 };
 
 type InputPillRenderingDetails = {
+    user_group_id?: number;
     display_value: string;
     has_image: boolean;
     img_src?: string;
@@ -58,6 +62,7 @@ type InputPillRenderingDetails = {
     has_status?: boolean;
     status_emoji_info?: EmojiRenderingDetails & {emoji_alt_code?: boolean};
     should_add_guest_user_indicator?: boolean;
+    group_size?: number;
 };
 
 // These are the functions that are exposed to other modules.
@@ -141,6 +146,7 @@ export function create<T>(opts: InputPillCreateOptions<T>): InputPillContainer<T
             const has_image = item.img_src !== undefined;
 
             const opts: InputPillRenderingDetails = {
+                user_group_id: item.group_id,
                 display_value: item.display_value,
                 has_image,
                 deactivated: item.deactivated,
@@ -157,6 +163,10 @@ export function create<T>(opts: InputPillCreateOptions<T>): InputPillContainer<T
                     opts.status_emoji_info = item.status_emoji_info;
                 }
                 opts.has_status = has_status;
+            }
+
+            if (store.pill_config?.show_user_group_size && item.group_size !== undefined) {
+                opts.group_size = item.group_size;
             }
 
             const pill_html = render_input_pill(opts);